project-root/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg[binary]==3.2.3
celery[redis]==5.4.0
redis==5.2.0
gunicorn==23.0.0

